// packages/components/sections/features/src/types.ts
export interface Images {
  image?: string;
  _key?: string;
  _type?: string;
  alt?: string;
}
// ...
export interface ArrayOfImageTitleAndText {
  mainImage?: {
    alt?: string;
    image?: string;
  };
  plainText?: string;
  title?: string;
  _key?: string;
  _type?: string;
}